requirements.txt
gradio>=4.44.1
python-dotenv==1.0.0
huggingface-hub>=0.17.0,<1.0
pytest>=8.0
cryptography>=41.0.0
pyyaml>=6.0
google-genai>=1.0.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-google-genai>=0.0.5
